# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music SNS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  bio: string | null;
  createdAt: string;
}

export interface UserProfile {
  id: number;
  username: string;
  /** @nullable */
  bio: string | null;
  postCount: number;
  createdAt: string;
}

export type PostSourceType = typeof PostSourceType[keyof typeof PostSourceType];


export const PostSourceType = {
  spotify: 'spotify',
  youtube: 'youtube',
  manual: 'manual',
} as const;

export interface Post {
  id: number;
  userId: number;
  username: string;
  songTitle: string;
  artistName: string;
  sourceType: PostSourceType;
  /** @nullable */
  sourceUrl: string | null;
  /** @nullable */
  message: string | null;
  /** @nullable */
  moodTag: string | null;
  likesCount: number;
  likedByMe: boolean;
  createdAt: string;
}

export interface LikeResult {
  likesCount: number;
  likedByMe: boolean;
}

export interface TrendingTag {
  tag: string;
  count: number;
}

export interface UserInput {
  /**
     * @minLength 2
     * @maxLength 30
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type PostInputSourceType = typeof PostInputSourceType[keyof typeof PostInputSourceType];


export const PostInputSourceType = {
  spotify: 'spotify',
  youtube: 'youtube',
  manual: 'manual',
} as const;

export interface PostInput {
  /** @minLength 1 */
  songTitle: string;
  /** @minLength 1 */
  artistName: string;
  sourceType: PostInputSourceType;
  sourceUrl?: string;
  /** @maxLength 140 */
  message?: string;
  moodTag?: string;
}

export interface BioUpdate {
  /** @maxLength 200 */
  bio: string;
}

export type ListPostsParams = {
limit?: number;
offset?: number;
};