# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Music SNS API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: posts
    description: Post operations
  - name: users
    description: User operations
  - name: stats
    description: Stats operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "204":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: Get timeline posts (newest first)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /posts/{id}:
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Post deleted
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [posts]
      summary: Like a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: unlikePost
      tags: [posts]
      summary: Unlike a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /users/{username}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /users/{username}/bio:
    patch:
      operationId: updateBio
      tags: [users]
      summary: Update user bio
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BioUpdate"
      responses:
        "200":
          description: Bio updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /users/{username}/posts:
    get:
      operationId: getUserPosts
      tags: [users]
      summary: Get posts by user
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User's posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"

  /stats/trending-tags:
    get:
      operationId: getTrendingTags
      tags: [stats]
      summary: Get trending mood tags
      responses:
        "200":
          description: Trending tags with counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrendingTag"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - username
        - bio
        - createdAt

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        bio:
          type: ["string", "null"]
        postCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - username
        - bio
        - postCount
        - createdAt

    Post:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        songTitle:
          type: string
        artistName:
          type: string
        sourceType:
          type: string
          enum: [spotify, youtube, manual]
        sourceUrl:
          type: ["string", "null"]
        message:
          type: ["string", "null"]
        moodTag:
          type: ["string", "null"]
        likesCount:
          type: integer
        likedByMe:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - username
        - songTitle
        - artistName
        - sourceType
        - sourceUrl
        - message
        - moodTag
        - likesCount
        - likedByMe
        - createdAt

    LikeResult:
      type: object
      properties:
        likesCount:
          type: integer
        likedByMe:
          type: boolean
      required:
        - likesCount
        - likedByMe

    TrendingTag:
      type: object
      properties:
        tag:
          type: string
        count:
          type: integer
      required:
        - tag
        - count

    UserInput:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 30
        password:
          type: string
          minLength: 8
          maxLength: 128
      required:
        - username
        - password

    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    PostInput:
      type: object
      properties:
        songTitle:
          type: string
          minLength: 1
        artistName:
          type: string
          minLength: 1
        sourceType:
          type: string
          enum: [spotify, youtube, manual]
        sourceUrl:
          type: string
        message:
          type: string
          maxLength: 140
        moodTag:
          type: string
      required:
        - songTitle
        - artistName
        - sourceType

    BioUpdate:
      type: object
      properties:
        bio:
          type: string
          maxLength: 200
      required:
        - bio