# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Music SNS API
 * OpenAPI spec version: 0.1.0
 */

export interface UserInput {
  /**
     * @minLength 2
     * @maxLength 30
     */
  username: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}